+++ src/main/java/br/com/pessoa/api/pessoa/PessoaService.java
@@ -4,7 +4,7 @@ import br.com.pessoa.api.core.CrudService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-public interface PessoaService extends CrudService<Pessoa, Long> {
+public interface PessoaService extends CrudService<Pessoa, Integer> {
 
     Page<PessoaDto> findAll(PessoaFiltro filtro, Pageable pageable);
 }
